+++ example/generateToken.php
@@ -1,0 +1,15 @@
+<?php
+
+use Ekopras18\Satusehat\Fhir;
+class generateToken
+{
+    public function __construct()
+    {
+        $this->fhir = new Fhir();
+    }
+
+    public function index()
+    {
+        return $this->fhir->token();
+    }
+}