+++ tests/Test.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Satusehat\Integration\Tests;
+
+use PHPUnit\Framework\TestCase;
+
+class Test extends TestCase
+{
+    public function testExample()
+    {
+        return 'coming soon...';
+    }
+}